+++ packages/server/src/__tests__/iterableEventEmitter.ts
@@ -1,0 +1,15 @@
+import EventEmitter from 'events';
+import { on } from 'node:events';
+
+type EventMap<T> = Record<keyof T, any[]>;
+
+export class IterableEventEmitter<
+  T extends EventMap<T>,
+> extends EventEmitter<T> {
+  toIterable<TEventName extends keyof T & string>(
+    eventName: TEventName,
+    opts?: NonNullable<Parameters<typeof on>[2]>,
+  ): AsyncIterable<T[TEventName]> {
+    return on(this as any, eventName, opts) as any;
+  }
+}
