+++ packages/tests/server/streaming.test.ts
@@ -1,8 +1,5 @@
 import { EventEmitter } from 'node:events';
-import {
-  routerToServerAndClientNew,
-  waitTRPCClientError,
-} from './___testHelpers';
+import { routerToServerAndClientNew } from './___testHelpers';
 import { waitError } from '@trpc/server/__tests__/waitError';
 import { waitFor } from '@testing-library/react';
 import type { TRPCLink } from '@trpc/client';
@@ -20,10 +17,17 @@ import {
   isAsyncIterable,
   run,
 } from '@trpc/server/unstable-core-do-not-import';
+import type { InferrableClientTypes } from '@trpc/server/unstable-core-do-not-import';
 import { konn } from 'konn';
 import superjson from 'superjson';
 import { z } from 'zod';
 import { zAsyncIterable } from './zAsyncIterable';
+
+async function waitTRPCClientError<TRoot extends InferrableClientTypes>(
+  fnOrPromise: Promise<unknown> | (() => unknown),
+) {
+  return waitError<TRPCClientError<TRoot>>(fnOrPromise, TRPCClientError);
+}
 
 describe('no transformer', () => {
   const orderedResults: number[] = [];
